Separated ViewController and ViewModel in SearchTableViewController file.
